Split large Rust modules into focused submodules
